+++ app/src/main/java/com/training/cinematic/activity/main/MainActivity.java
@@ -16,7 +16,7 @@ import android.widget.Toast;
 import com.training.cinematic.Adapter.MainAdapter;
 import com.training.cinematic.Fragment.poplarmovie.PopularMoviesFragment;
 import com.training.cinematic.Fragment.populartv.PopularTvFragment;
-import com.training.cinematic.Fragment.UpComingMovieFragment;
+import com.training.cinematic.Fragment.upcomingmovie.UpComingMovieFragment;
 import com.training.cinematic.R;
 import com.training.cinematic.Utils.SharedPrefsHelp;
 import com.training.cinematic.activity.Profile.ProfileActivity;
